Split up AddFunction.h into individual header files
