--- src/lib/index.h
@@ -1,8 +1,0 @@
-#ifndef NODE_RAYLIB_LIB_INDEX_H_
-#define NODE_RAYLIB_LIB_INDEX_H_
-
-#include "AddFunction.h"
-#include "ToObject.h"
-#include "ValidArgs.h"
-
-#endif
