+++ src/timing.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <napi.h>
 #include <raylib.h>
-#include "lib/index.h"
+#include "lib/AddFunction.h"
 
 void node_raylib_init_timing(Napi::Env& env, Napi::Object& exports) {
   AddFunction(env, exports, "SetTargetFPS", SetTargetFPS);
